Skni6e9M/3 start working with intermediate results (radio controls) storing a new field is_intermediate in splits

+++ app_rest/plugins/Results/src/Model/Table/SplitsTable.php
@@ -61,6 +61,7 @@ class SplitsTable extends AppTable
     {
         if ($splits) {
             foreach ($splits as $split) {
+                $split['is_intermediate'] = $helper->getChecker()->isIntermediates();
                 $splitToSave = $this->fillNewWithStage($split, $helper->getEventId(), $helper->getStageId());
                 if ($split['station'] ?? null) {
                     $control = $this->Controls->createControlIfNotExists($helper, $split);
@@ -78,7 +79,9 @@ class SplitsTable extends AppTable
         $helper->getMetrics()->startSplitsTime();
         if ($splits && !$runnerResultToSave->hasSameSplits($splits)) {
             $runnerResultToSave->setHash($splits);
-            $this->deleteAllByRunnerId($runnerResultToSave->id);
+            if (!$helper->getChecker()->isIntermediates()) {
+                $this->deleteAllByRunnerId($runnerResultToSave->id);
+            }
             $runnerResultToSave = $this->uploadForEachSplit($runnerResultToSave, $splits, $helper);
         }
         $helper->getMetrics()->endSplitsTime();
